fix 1

+++ Simple_File_System/OS_File_System.c
@@ -114,7 +114,7 @@ uint8_t OS_File_Append(uint8_t num, uint8_t buf[512]) {
 }
 
 uint8_t OS_File_Read(uint8_t num, uint8_t location, uint8_t buf[512]) {
-    uint8_t i;
+    uint16_t i;
     uint8_t sector;
     uint32_t addr;
     uint8_t *flashPtr;
